+++ src/curvesimilarities/__init__.py
@@ -1,6 +1,6 @@
 """Curve similarity measures."""
 
-from .dtw import dtw, dtw_owp, sdtw, sdtw_owp
+from .dtw import dtw, dtw_owp
 from .frechet import dfd, fd
 from .integfrechet import ifd, ifd_owp
 
@@ -9,8 +9,6 @@ __all__ = [
     "dfd",
     "dtw",
     "dtw_owp",
-    "sdtw",
-    "sdtw_owp",
     "ifd",
     "ifd_owp",
 ]